pyproject.toml
[project]
name = "email-assistant"
version = "0.1.0"
description = "Personalized assistant for email responses"
authors = [
    {name = "Sofia Samoylova",email = "samoilova.se@phystech.edu"}
]
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "torch (==2.2.1)",
    "tqdm (>=4.67.1,<5.0.0)",
    "transformers (==4.51.3)",
    "bitsandbytes (==0.43.2)",
    "accelerate (==1.6.0)",
    "numpy (==1.26.4)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pytorch-lightning (>=2.5.1.post0,<3.0.0)",
    "peft (>=0.15.2,<0.16.0)",
]

[tool.poetry]
packages = [{include = "email_assistant", from = "src"}]

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
hydra-core = "^1.3.2"
dvc = "^3.59.2"
mlflow = "^2.22.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
